restructure examples/ directory based on model language
