--- examples/bayesian_linear_regression_placeholder.py
@@ -1,57 +1,0 @@
-#!/usr/bin/env python
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-import edward as ed
-import numpy as np
-import tensorflow as tf
-
-from edward.models import Normal
-from scipy.stats import norm
-
-
-def build_toy_dataset(N, noise_std=0.1):
-  X = np.concatenate([np.linspace(0, 2, num=N / 2),
-                      np.linspace(6, 8, num=N / 2)])
-  y = 5.0 * X + norm.rvs(0, noise_std, size=N)
-  X = X.reshape((N, 1))
-  return X.astype(np.float32), y.astype(np.float32)
-
-
-ed.set_seed(42)
-
-N = 40  # num data points
-D = 1  # num features
-
-# DATA
-X_train, y_train = build_toy_dataset(N)
-X_test, y_test = build_toy_dataset(N)
-
-# MODEL
-X = ed.placeholder(tf.float32, [N, D])
-w = Normal(mu=tf.zeros(D), sigma=tf.ones(D))
-b = Normal(mu=tf.zeros(1), sigma=tf.ones(1))
-y = Normal(mu=ed.dot(X, w) + b, sigma=tf.ones(N))
-
-# INFERENCE
-qw = Normal(mu=tf.Variable(tf.random_normal([D])),
-            sigma=tf.nn.softplus(tf.Variable(tf.random_normal([D]))))
-qb = Normal(mu=tf.Variable(tf.random_normal([1])),
-            sigma=tf.nn.softplus(tf.Variable(tf.random_normal([1]))))
-
-data = {X: X_train, y: y_train}
-inference = ed.MFVI({w: qw, b: qb}, data)
-inference.initialize()
-
-sess = ed.get_session()
-for t in range(1001):
-  _, loss = sess.run([inference.train, inference.loss], {X: data[X]})
-  inference.print_progress(t, loss)
-
-# CRITICISM
-y_post = ed.copy(y, {w: qw.mean(), b: qb.mean()})
-# This is equivalent to
-# y_post = Normal(mu=ed.dot(X, qw.mean()) + qb.mean(), sigma=tf.ones(N))
-
-print(ed.evaluate('mean_squared_error', data={X: X_test, y_post: y_test}))
