+++ examples/bayesian_linear_regression_tensorboard.py
