+++ examples/bayesian_nn.py
@@ -16,74 +16,12 @@ from __future__ import division
 from __future__ import print_function
 
 import edward as ed
-import matplotlib.pyplot as plt
 import numpy as np
 import tensorflow as tf
 
 from edward.models import Normal
 from edward.stats import norm
 from edward.util import rbf
-
-
-class BayesianNN:
-  """
-  Bayesian neural network for regressing outputs y on inputs x.
-
-  p((x,y), z) = Normal(y | NN(x; z), lik_std) *
-                Normal(z | 0, prior_std),
-
-  where z are neural network weights, and with known likelihood and
-  prior standard deviations.
-
-  Parameters
-  ----------
-  layer_sizes : list
-    The size of each layer, ordered from input to output.
-  nonlinearity : function, optional
-    Non-linearity after each linear transformation in the neural
-    network; aka activation function.
-  lik_std : float, optional
-    Standard deviation of the normal likelihood; aka noise parameter,
-    homoscedastic noise, scale parameter.
-  prior_std : float, optional
-    Standard deviation of the normal prior on weights; aka L2
-    regularization parameter, ridge penalty, scale parameter.
-  """
-  def __init__(self, layer_sizes, nonlinearity=tf.nn.tanh,
-               lik_std=0.1, prior_std=1.0):
-    self.layer_sizes = layer_sizes
-    self.nonlinearity = nonlinearity
-    self.lik_std = lik_std
-    self.prior_std = prior_std
-
-    self.n_layers = len(layer_sizes)
-    self.weight_dims = list(zip(layer_sizes[:-1], layer_sizes[1:]))
-    self.n_vars = sum((m + 1) * n for m, n in self.weight_dims)
-
-  def unpack_weights(self, zs):
-    """Unpack weight matrices and biases from a flattened vector."""
-    for m, n in self.weight_dims:
-      yield tf.reshape(zs[:(m * n)], [m, n]), \
-          tf.reshape(zs[(m * n):(m * n + n)], [n])
-      zs = zs[(m + 1) * n:]
-
-  def neural_network(self, x, zs):
-    """Forward pass of the neural net, outputting a vector of
-    `n_minibatch` elements."""
-    h = x
-    for W, b in self.unpack_weights(zs):
-      h = self.nonlinearity(tf.matmul(h, W) + b)
-
-    return tf.squeeze(h)  # n_minibatch x 1 to n_minibatch
-
-  def log_prob(self, xs, zs):
-    """Return scalar, the log joint density log p(xs, zs)."""
-    x, y = xs['x'], xs['y']
-    zs = zs['z']
-    log_prior = tf.reduce_sum(norm.logpdf(zs, 0.0, self.prior_std))
-    mu = self.neural_network(x, zs)
-    log_lik = tf.reduce_sum(norm.logpdf(y, mu, self.lik_std))
-    return log_lik + log_prior
 
 
 def build_toy_dataset(N=40, noise_std=0.1):
@@ -96,51 +34,53 @@ def build_toy_dataset(N=40, noise_std=0.1):
   return x, y
 
 
+def neural_network(x):
+    h = tf.nn.tanh(tf.matmul(x, W_0) + b_0)
+    h = tf.nn.tanh(tf.matmul(h, W_1) + b_1)
+    h = tf.matmul(h, W_2) + b_2
+    return tf.reshape(h, [-1])
+
+
 ed.set_seed(42)
-x_train, y_train = build_toy_dataset()
 
-model = BayesianNN(layer_sizes=[1, 10, 10, 1], nonlinearity=rbf)
+N = 40  # num data ponts
+D = 1   # num features
 
-qz_mu = tf.Variable(tf.random_normal([model.n_vars]))
-qz_sigma = tf.nn.softplus(tf.Variable(tf.random_normal([model.n_vars])))
-qz = Normal(mu=qz_mu, sigma=qz_sigma)
+# DATA
+x_train, y_train = build_toy_dataset(N)
 
-# Set up figure
-fig = plt.figure(figsize=(8, 8), facecolor='white')
-ax = fig.add_subplot(111, frameon=False)
-plt.ion()
-plt.show(block=False)
+# MODEL
+W_0 = Normal(mu=tf.zeros([D, 10]), sigma=tf.ones([D, 10]))
+W_1 = Normal(mu=tf.zeros([10, 10]), sigma=tf.ones([10, 10]))
+W_2 = Normal(mu=tf.zeros([10, 1]), sigma=tf.ones([10, 1]))
+b_0 = Normal(mu=tf.zeros(10), sigma=tf.ones(10))
+b_1 = Normal(mu=tf.zeros(10), sigma=tf.ones(10))
+b_2 = Normal(mu=tf.zeros(1), sigma=tf.ones(1))
 
-sess = ed.get_session()
-data = {'x': x_train, 'y': y_train}
-inference = ed.MFVI({'z': qz}, data, model)
+x = tf.convert_to_tensor(x_train, dtype=tf.float32)
+y = Normal(mu=neural_network(x), sigma=0.1 * tf.ones(N))
+
+# INFERENCE
+qW_0 = Normal(mu=tf.Variable(tf.random_normal([D, 10])),
+              sigma=tf.nn.softplus(tf.Variable(tf.random_normal([D, 10]))))
+qW_1 = Normal(mu=tf.Variable(tf.random_normal([10, 10])),
+              sigma=tf.nn.softplus(tf.Variable(tf.random_normal([10, 10]))))
+qW_2 = Normal(mu=tf.Variable(tf.random_normal([10, 1])),
+              sigma=tf.nn.softplus(tf.Variable(tf.random_normal([10, 1]))))
+qb_0 = Normal(mu=tf.Variable(tf.random_normal([10])),
+              sigma=tf.nn.softplus(tf.Variable(tf.random_normal([10]))))
+qb_1 = Normal(mu=tf.Variable(tf.random_normal([10])),
+              sigma=tf.nn.softplus(tf.Variable(tf.random_normal([10]))))
+qb_2 = Normal(mu=tf.Variable(tf.random_normal([1])),
+              sigma=tf.nn.softplus(tf.Variable(tf.random_normal([1]))))
+
+data = {y: y_train}
+inference = ed.MFVI({W_0: qW_0, b_0: qb_0,
+                     W_1: qW_1, b_1: qb_1,
+                     W_2: qW_2, b_2: qb_2}, data)
+
 inference.initialize(n_print=100)
 for t in range(1000):
   loss = inference.update()
   if t % inference.n_print == 0:
     print("iter {:d} loss {:.2f}".format(t, loss))
-
-    # Sample functions from variational model
-    mean, std = sess.run([qz.mu, qz.sigma])
-    rs = np.random.RandomState(0)
-    zs = rs.randn(10, model.n_vars) * std + mean
-    zs = tf.convert_to_tensor(zs, dtype=tf.float32)
-    inputs = np.linspace(-8, 8, num=400, dtype=np.float32)
-    x = tf.expand_dims(inputs, 1)
-    mus = []
-    for z in tf.unpack(zs):
-      mus += [model.neural_network(x, z)]
-
-    outputs = tf.pack(mus).eval()
-
-    # Get data
-    x, y = data['x'], data['y']
-
-    # Plot data and functions
-    plt.cla()
-    ax.plot(x, y, 'bx')
-    ax.plot(inputs, outputs.T)
-    ax.set_xlim([-8, 8])
-    ax.set_ylim([-2, 3])
-    plt.draw()
-    plt.pause(1.0 / 60.0)
