+++ examples/beta_bernoulli.py
