+++ examples/beta_bernoulli_map.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python
 """
-A simple coin flipping example. The model is written in TensorFlow.
-Inspired by Stan's toy example.
+A simple coin flipping example. Inspired by Stan's toy example.
 
 Probability model
   Prior: Beta
   Likelihood: Bernoulli
-Inference: Maximum a posteriori
+Variational model
+  Likelihood: Mean-field Beta
 """
 from __future__ import absolute_import
 from __future__ import division
@@ -16,25 +16,17 @@ import edward as ed
 import numpy as np
 import tensorflow as tf
 
-from edward.models import Beta
-from edward.stats import bernoulli, beta
-
-
-class BetaBernoulli:
-  """p(x, p) = Bernoulli(x | p) * Beta(p | 1, 1)"""
-  def __init__(self):
-    self.n_vars = 1
-
-  def log_prob(self, xs, zs):
-    log_prior = beta.logpdf(zs['p'], a=1.0, b=1.0)
-    log_lik = tf.reduce_sum(bernoulli.logpmf(xs['x'], p=zs['p']))
-    return log_lik + log_prior
-
+from edward.models import Bernoulli, Beta, PointMass
 
 ed.set_seed(42)
-data = {'x': np.array([0, 1, 0, 0, 0, 0, 0, 0, 0, 1])}
 
-model = BetaBernoulli()
+p = Beta(a=1.0, b=1.0)
+x = Bernoulli(p=tf.ones(10) * p)
 
-inference = ed.MAP(['p'], data, model)
-inference.run(n_iter=100, n_print=10)
+data = {x: np.array([0, 1, 0, 0, 0, 0, 0, 0, 0, 1])}
+
+qp_params = tf.nn.sigmoid(tf.Variable(tf.random_normal([])))
+qp = PointMass(params=qp_params)
+
+inference = ed.MAP({p: qp}, data)
+inference.run(n_iter=50, n_print=10)
