+++ examples/np_beta_bernoulli.py
