+++ examples/pymc3_beta_bernoulli.py
