+++ examples/stan_beta_bernoulli.py
