+++ examples/tf_bayesian_linear_regression_plot.py
