+++ examples/tf_bayesian_nn_analytic_kl.py
