+++ examples/tf_bayesian_nn_separate_weights.py
