+++ examples/tf_bernoulli.py
