+++ examples/tf_beta_bernoulli.py
