+++ examples/tf_beta_bernoulli_prior_predictive_check.py
