+++ examples/tf_convolutional_vae_util.py
