+++ examples/tf_hierarchical_logistic_regression.py
