+++ examples/tf_latent_space_model.py
