+++ examples/tf_mixture_density_network_demo.py
