+++ examples/tf_mixture_gaussian.py
