+++ examples/tf_mixture_gaussian_laplace.py
