+++ examples/tf_mixture_gaussian_map.py
