+++ examples/tf_normal_idiomatic_tf.py
